Make GetUnitEquippedWeapon return u32 instead of u16

+++ src/bmitem.c
@@ -476,7 +476,7 @@ u16 GetItemAfterUse(int item) {
     return item; // return used item
 }
 
-u16 GetUnitEquippedWeapon(struct Unit* unit) {
+u32 GetUnitEquippedWeapon(struct Unit* unit) {
     int i;
 
     for (i = 0; i < UNIT_ITEM_COUNT; ++i)
